pause before lunch

+++ src/lib.rs
@@ -1,6 +1,6 @@
 #![allow(dead_code)]
 
-pub type Move = [u8; 2];
+pub type Move = [usize; 2];
 pub fn add(left: usize, right: usize) -> usize {
     left + right
 }
